requirements.txt
pymysql
mysql-connector-python
PyMySQL
cryptography